Fixed liveliness by switching to the next lottery winner after each epoch. Added epoch to configuration file

+++ src/main/java/configuration/Configuration.kt
@@ -9,7 +9,7 @@ data class Configuration(
         val listeningPort: Int,
         val maxNodes: Int,
         val keystorePath : String,
-        val epochDuration : Int,
+        val epochDuration : Long,
         val broadcastSpread : Int
 ) {
     val trustedHttpAddress: String get() = "http://$trustedNodeIP:$trustedNodePort"
